Moves response-status-bar to it's own directory.
